==> notebook/03_encoding_standarization_train.ipynb <==
import pandas as pd
from sklearn.preprocessing import RobustScaler
from sklearn.preprocessing import StandardScaler

df = pd.read_csv('../data/train_data_wo_outliers.csv')
df

#encoding categorical variables
mapping = {'quality':['Fair', 'Good', 'Very Good', 'Premium', 'Ideal'], 
            'color':['J', 'I', 'H', 'G', 'F', 'E', 'D'],
            'clarity':['I1', 'SI2', 'SI1', 'VS2', 'VS1', 'VVS2', 'VVS1', 'IF']}
for k,v in mapping.items():
    for i in v:
        df[k] = df[k].replace(i, v.index(i))
df

#dummies for nominal variables
#dummies = pd.get_dummies(df[['quality', 'color', 'clarity']], prefix_sep = "_", prefix = ['quality_EN', 'color_EN', 'clarity_EN'], dtype = int)
#df = pd.concat([df, dummies], axis = 1).drop(['quality', 'color', 'clarity'], axis = 1)
#df

#standardizing the data robust
robust = RobustScaler()
robust.fit(df[["weight", "depth_percent", "table_percent", "volume"]])
X_robust = robust.transform(df[["weight", "depth_percent", "table_percent", "volume"]])
df[["weight_ES", "depth_percent_ES", "table_percent_ES", "volume_ES"]] = X_robust
df.drop(["weight", "depth_percent", "table_percent", "volume"], axis=1, inplace=True)
df.head()

#standardizing the data standard
#scaler = StandardScaler()
#scaler.fit(df[["weight", "depth_percent", "table_percent", "volume"]])
#X_scaled = scaler.transform(df[["weight", "depth_percent", "table_percent", "volume"]])
#df[["weight_ES", "depth_percent_ES", "table_percent_ES", "volume_ES"]] = X_scaled
#df.drop(["weight", "depth_percent", "table_percent", "volume"], axis=1, inplace=True)
#df.head()

df.to_csv('../data/train_data_encoded.csv', index=False)
