==> notebook/04_model_1_dt.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV
from sklearn.tree import DecisionTreeRegressor
from sklearn import tree
import matplotlib.pyplot as plt
from sklearn import metrics

df = pd.read_csv('../data/train_data_encoded.csv')
print('shape:', df.shape)
df.head(5)

X = df.drop(["price"], axis = 1)
y = df['price']

X_train, X_test, y_train, y_test = train_test_split(X, y, train_size   = 0.8, random_state = 42)

print("Datos de entrenamiento")
print("-----------------------")
print(y_train.describe())
print('\n')
print("Datos de testeo")
print("-----------------------")
print(y_test.describe())

def metricas(y_test, y_train, y_test_pred, y_train_pred, tipo_modelo):
    resultados = {'MAE': [metrics.mean_absolute_error(y_test, y_test_pred), metrics.mean_absolute_error(y_train, y_train_pred)],
                'MSE': [metrics.mean_squared_error(y_test, y_test_pred), metrics.mean_squared_error(y_train, y_train_pred)],
                'RMSE': [np.sqrt(metrics.mean_squared_error(y_test, y_test_pred)), np.sqrt(metrics.mean_squared_error(y_train, y_train_pred))],
                'R2':  [metrics.r2_score(y_test, y_test_pred), metrics.r2_score(y_train, y_train_pred)],
                 "set": ["test", "train"]}
    df = pd.DataFrame(resultados)
    df["modelo"] = tipo_modelo
    return df

param = {"max_depth": [26, 32, 40],
        "max_features": [7, 9],
        "min_samples_split": [75, 100]}

decision_tree = GridSearchCV(
            estimator=DecisionTreeRegressor(),
            param_grid= param,
            cv=10,
            n_jobs=-1,
            verbose=3,
            return_train_score = True,
            scoring="neg_mean_squared_error")

%time
decision_tree.fit(X_train, y_train)
best_dt = decision_tree.best_estimator_
print('best tree:', best_dt)

y_pred_test_dt = best_dt.predict(X_test)
y_pred_train_dt = best_dt.predict(X_train)
dt_results_1 = metricas(y_test, y_train, y_pred_test_dt, y_pred_train_dt, "Decision tree 1")
dt_results_1.style.background_gradient(cmap='coolwarm')

param = {"max_depth": [8, 10, 12],
        "max_features": [4, 8],
        "min_samples_split": [10, 50, 100]}

decision_tree = GridSearchCV(
            estimator=DecisionTreeRegressor(),
            param_grid= param,
            cv=10,
            n_jobs=-1,
            verbose=1,
            return_train_score = True,
            scoring="neg_mean_squared_error")

%time
decision_tree.fit(X_train, y_train)
best_dt = decision_tree.best_estimator_
print('best tree:', best_dt)
y_pred_test_dt = best_dt.predict(X_test)
y_pred_train_dt = best_dt.predict(X_train)
dt_results_2 = metricas(y_test, y_train, y_pred_test_dt, y_pred_train_dt, "Decision tree 2")
dt_results_2.style.background_gradient(cmap='coolwarm')

param = {"max_depth": [12, 18],
        "max_features": [8, 12],
        "min_samples_split": [10, 50, 100]}

decision_tree = GridSearchCV(
            estimator=DecisionTreeRegressor(),
            param_grid= param,
            cv=10,
            n_jobs=-1,
            verbose=1,
            return_train_score = True,
            scoring="neg_mean_squared_error")

%time
decision_tree.fit(X_train, y_train)
best_dt = decision_tree.best_estimator_
print('best tree:', best_dt)
y_pred_test_dt = best_dt.predict(X_test)
y_pred_train_dt = best_dt.predict(X_train)
dt_results_3 = metricas(y_test, y_train, y_pred_test_dt, y_pred_train_dt, "Decision tree 3")
dt_results_3.style.background_gradient(cmap='coolwarm')

df_results = pd.concat([dt_results_1, dt_results_2, dt_results_3], axis = 0)
df_results
