==> notebook/04_model_1_dt_importance.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.tree import DecisionTreeRegressor
import statsmodels.formula.api as smf

df = pd.read_csv('../data/train_data_encoded.csv')
print('shape:', df.shape)
df.head(5)

#results using stats library
results = smf.ols("price ~  quality + color + clarity + weight_ES + depth_percent_ES + table_percent_ES + volume_ES", data=df).fit() 
results.summary()

#split dara train and test
X = df.drop(["price"], axis = 1)
y = df['price']

X_train, X_test, y_train, y_test = train_test_split(X, y, train_size   = 0.8, random_state = 42)

print("Datos de entrenamiento")
print("-----------------------")
print(y_train.describe())
print('\n')
print("Datos de testeo")
print("-----------------------")
print(y_test.describe())

X.sample(5)

pd.DataFrame(y).sample(5)

#predictor variables importance model
regressor = DecisionTreeRegressor(random_state = 0) 
regressor.fit(X_train, y_train)
max_features = np.sqrt(len(X_train.columns))
print('max_features:', max_features)
print('max_depth:', regressor.tree_.max_depth)
y_pred_test_dt = regressor.predict(X_test)
y_pred_train_dt = regressor.predict(X_train)

predictor_importance = pd.DataFrame({'predictor': X_train.columns, 'importance': regressor.feature_importances_})

predictor_importance.sort_values( ascending=False, by= "importance", inplace=True)
print("predictor variables importance")
print("-------------------------------------------")
predictor_importance

#average price per weight range and color
df['weight_range'] = df['weight_ES'].round(0).astype('category')
weight_clarity_analysis = pd.DataFrame(df.groupby(['weight_range', 'clarity']).mean().head(50).sort_index(ascending=False)['price']).dropna()
weight_clarity_analysis
