==> notebook/04_model_2_rf.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import GridSearchCV
from sklearn import metrics

df = pd.read_csv('../data/train_data_encoded.csv')
print('shape:', df.shape)
df.head(5)

#split dara train and test
X = df.drop(["price"], axis = 1)
y = df['price']

X_train, X_test, y_train, y_test = train_test_split(X, y, train_size   = 0.8, random_state = 42)

print("Datos de entrenamiento")
print("-----------------------")
print(y_train.describe())
print('\n')
print("Datos de testeo")
print("-----------------------")
print(y_test.describe())

def metricas(y_test, y_train, y_test_pred, y_train_pred, tipo_modelo):
    resultados = {'MAE': [metrics.mean_absolute_error(y_test, y_test_pred), metrics.mean_absolute_error(y_train, y_train_pred)],
                'MSE': [metrics.mean_squared_error(y_test, y_test_pred), metrics.mean_squared_error(y_train, y_train_pred)],
                'RMSE': [np.sqrt(metrics.mean_squared_error(y_test, y_test_pred)), np.sqrt(metrics.mean_squared_error(y_train, y_train_pred))],
                'R2':  [metrics.r2_score(y_test, y_test_pred), metrics.r2_score(y_train, y_train_pred)],
                 "set": ["test", "train"]}
    df = pd.DataFrame(resultados)
    df["modelo"] = tipo_modelo
    return df

param = {"max_depth": [18],
        "max_features": [15],
        "min_samples_split": [10]}

random_forest = GridSearchCV(
            estimator=RandomForestRegressor(),
            param_grid= param,
            cv=10,
            n_jobs = -1,
            verbose= 3,
            return_train_score = True,
            scoring="neg_mean_squared_error")

random_forest.fit(X_train, y_train)

best_rf = random_forest.best_estimator_
print(best_rf)

y_pred_test_rf = random_forest.predict(X_test)
y_pred_train_rf = random_forest.predict(X_train)
rf_results_1 = metricas(y_test, y_train, y_pred_test_rf, y_pred_train_rf, "Random Forest 1")
rf_results_1.style.background_gradient(cmap='coolwarm')

param = {"max_depth": [15, 18, 22],
        "max_features": [12, 15, 18],
        "min_samples_split": [5, 10, 15]}

random_forest = GridSearchCV(
            estimator=RandomForestRegressor(),
            param_grid= param,
            cv=10,
            n_jobs = -1,
            verbose= 1,
            return_train_score = True,
            scoring="neg_mean_squared_error")

random_forest.fit(X_train, y_train)

best_rf = random_forest.best_estimator_
print(best_rf)

y_pred_test_rf = random_forest.predict(X_test)
y_pred_train_rf = random_forest.predict(X_train)
rf_results_2 = metricas(y_test, y_train, y_pred_test_rf, y_pred_train_rf, "Random Forest 2")
rf_results_2.style.background_gradient(cmap='coolwarm')

param = {"max_depth": [20],
        "max_features": [17],
        "min_samples_split": [12]}

random_forest = GridSearchCV(
            estimator=RandomForestRegressor(),
            param_grid= param,
            cv=10,
            n_jobs = -1,
            verbose= 1,
            return_train_score = True,
            scoring="neg_mean_squared_error")

random_forest.fit(X_train, y_train)

best_rf = random_forest.best_estimator_
print(best_rf)

y_pred_test_rf = random_forest.predict(X_test)
y_pred_train_rf = random_forest.predict(X_train)
rf_results_3 = metricas(y_test, y_train, y_pred_test_rf, y_pred_train_rf, "Random Forest 3")
rf_results_3.style.background_gradient(cmap='coolwarm')

df_results = pd.concat([rf_results_1, rf_results_2, rf_results_3], axis = 0)
df_results
