==> notebook/05_predict_test_model_4_xgb.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV
from sklearn import metrics
import xgboost as xgb

df = pd.read_csv('../data/train_data_encoded.csv')
#df.price = np.log(df.price)
print('shape:', df.shape)
df.head(5)

X = df.drop(["price"], axis = 1)
y = df['price']

X_train, X_test, y_train, y_test = train_test_split(X, y, train_size   = 0.8, random_state = 42)

def metricas(y_test, y_train, y_test_pred, y_train_pred, tipo_modelo):
    resultados = {'MAE': [metrics.mean_absolute_error(y_test, y_test_pred), metrics.mean_absolute_error(y_train, y_train_pred)],
                'MSE': [metrics.mean_squared_error(y_test, y_test_pred), metrics.mean_squared_error(y_train, y_train_pred)],
                'RMSE': [np.sqrt(metrics.mean_squared_error(y_test, y_test_pred)), np.sqrt(metrics.mean_squared_error(y_train, y_train_pred))],
                'R2':  [metrics.r2_score(y_test, y_test_pred), metrics.r2_score(y_train, y_train_pred)],
                 "set": ["test", "train"]}
    df = pd.DataFrame(resultados)
    df["modelo"] = tipo_modelo
    return df

params = { 'max_depth': [7],
           'learning_rate': [0.02],
           'n_estimators': [750],
           'colsample_bytree': [0.9]}
xgbr = xgb.XGBRegressor(seed = 20, objective ='reg:squarederror')
clf = GridSearchCV(estimator=xgbr, 
                   param_grid=params,
                   scoring='neg_mean_squared_error', 
                   verbose=1,
                   n_jobs=-1)
clf.fit(X_train, y_train)

print("Best parameters:", clf.best_params_)
print("Lowest RMSE: ", (-clf.best_score_)**(1/2.0))

y_pred_gb_test = clf.predict(X_test)
y_pred_gb_train= clf.predict(X_train)

results_xg_boost = metricas(y_test, y_train, y_pred_gb_test, y_pred_gb_train, "XG Boost GridSearch")
results_xg_boost.style.background_gradient(cmap='coolwarm')

df_test = pd.read_csv('../data/test_data_encoded.csv')
df_test

#predict test to submit competition
y_pred_clf_test = clf.predict(df_test)
submission_4 = pd.DataFrame(y_pred_clf_test).reset_index()
print('shape:', y_pred_clf_test.shape)
submission_4.columns = ['id', 'price']
submission_4

submission_4.to_csv('../output/submission_4.csv', index=False)
