==> notebook/05_predict_test_model_5_ avg_models.ipynb <==
import pandas as pd

df1 = pd.read_csv('../output/submission_1.csv').drop(['id'], axis=1).rename(columns={'price': 'submission_1'})
df2 = pd.read_csv('../output/submission_2.csv').drop(['id'], axis=1).rename(columns={'price': 'submission_2'})
df3 = pd.read_csv('../output/submission_3.csv').drop(['id'], axis=1).rename(columns={'price': 'submission_3'})
df4 = pd.read_csv('../output/submission_4.csv').drop(['id'], axis=1).rename(columns={'price': 'submission_4'})

submission_5 = pd.concat([df1, df2, df3, df4], axis=1)
submission_5['price'] = submission_5.mean(axis=1)
submission_5.drop(['submission_1', 'submission_2', 'submission_3', 'submission_4'], axis=1, inplace=True)
submission_5.reset_index(drop=False, inplace=True)
submission_5.rename(columns={'index': 'id'}, inplace=True)

submission_5

submission_5.to_csv('../output/submission_5_avg.csv', index=False)
